+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* Use std string literals
+
 ## 0.6.0 &ndash; 2021-02-02
 
 ### Changed